reduced inference time per token from ~2.5 minutes to ~40 seconds on my MBP.
we no longer create a new WasmInstance on each token. this overhead slowed things down significantly. to permit the reuse of the component, we added  calls to tell the component to reset its state in preparation for the next call. still lots of refactoring to be done, but this was a necessary clean-up.